requirements.txt
faker
psycopg2-binary
pyyaml